+++ src/Settings/AccountTypeSettings/index.js
@@ -1,0 +1,1 @@
+export { default as AccountTypeSettings } from './AccountTypeSettings';