+++ src/Settings/BankingInformationSettings/constants.js
@@ -1,0 +1,4 @@
+export const typeStatus = [
+  { value: 'Active' },
+  { value: 'Inactive' },
+];