+++ src/Settings/BankingInformationSettings/index.js
@@ -1,0 +1,1 @@
+export { default as BankingInformationSettings } from './BankingInformationSettings';