+++ src/Settings/SettingsPage.js
@@ -5,6 +5,8 @@ import { Settings } from '@folio/stripes/smart-components';
 
 import { CategorySettings } from './CategorySettings';
 import { TypeSettings } from './TypeSettings';
+import { BankingInformationSettings } from './BankingInformationSettings';
+import { AccountTypeSettings } from './AccountTypeSettings';
 
 const pages = [
   {
@@ -18,6 +20,18 @@ const pages = [
     label: <FormattedMessage id="ui-organizations.settings.types" />,
     perm: 'settings.organizations.enabled',
     route: 'type',
+  },
+  {
+    component: BankingInformationSettings,
+    label: <FormattedMessage id="ui-organizations.settings.bankingInformation" />,
+    perm: 'settings.organizations.enabled',
+    route: 'banking-information',
+  },
+  {
+    component: AccountTypeSettings,
+    label: <FormattedMessage id="ui-organizations.settings.accountTypes" />,
+    perm: 'settings.organizations.enabled',
+    route: 'account-types',
   },
 ];